refactor(search): make experimental search component a regular one
BREAKING CHANGE: Search has new import path - carbon-react/lib/components/search
