+++ src/components/drawer/__internal__/drawer.stories.mdx
@@ -6,7 +6,7 @@ import { action } from '@storybook/addon-actions';
 import styled from 'styled-components';
 import Button from '../../button';
 import PopoverContainer from '../../popover-container';
-import Search from '../../../__experimental__/components/search';
+import Search from '../../search';
 import DialogFullScreen from '../../dialog-full-screen';
 import { StyledDrawerContent, StyledDrawerSidebar } from './drawer.style';
 import Heading from '../../heading';
