+++ src/components/menu/menu.stories.mdx
@@ -11,7 +11,7 @@ import Box from "../box";
 import NavigationBar from "../navigation-bar";
 import VerticalDivider from "../vertical-divider";
 import { SageIcon } from "../../../.storybook/welcome-page/footer/footer.style.js";
-import Search from "../../__experimental__/components/search";
+import Search from "../search";
 
 <Meta title="Design System/Menu" parameters={{ info: { disable: true } }} />
