+++ src/components/search/index.d.ts
