+++ src/components/search/search.component.js
@@ -6,12 +6,12 @@ import StyledSearch, {
   StyledSearchButton,
   StyledButtonIcon,
 } from "./search.style";
-import tagComponent from "../../../utils/helpers/tags";
-import { filterStyledSystemMarginProps } from "../../../style/utils";
-import Icon from "../../../components/icon";
-import Textbox from "../../../components/textbox";
-import Button from "../../../components/button";
-import Events from "../../../utils/helpers/events";
+import tagComponent from "../../utils/helpers/tags";
+import { filterStyledSystemMarginProps } from "../../style/utils";
+import Icon from "../icon";
+import Textbox from "../textbox";
+import Button from "../button";
+import Events from "../../utils/helpers/events";
 
 const marginPropTypes = filterStyledSystemMarginProps(
   styledSystemPropTypes.space
