+++ src/components/search/search.stories.mdx
@@ -1,9 +1,9 @@
 import { Meta, Story, Preview, Props} from '@storybook/addon-docs/blocks';
 import {useState} from 'react';
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import Search from '.';
-import Button from '../../../components/button';
-import Box from '../../../components/box';
+import Button from '../button';
+import Box from '../box';
 
 <Meta title="Design System/Search" />
 
@@ -19,7 +19,7 @@ This search component should be used if you require the user to conduct a search
 Import `Search` into the project.
 
 ```javascript
-import Search from 'carbon-react/lib/__experimental__/components/search';
+import Search from 'carbon-react/lib/components/search';
 ```
 
 ## Examples
@@ -28,7 +28,7 @@ import Search from 'carbon-react/lib/__experimental__/components/search';
 
 <Preview>
   <Story name="default" parameters={{ info: { disable: true }}} >
-    <Search placeholder="Search..." defaultValue=""/>
+    <Search placeholder="Search..." defaultValue="" />
   </Story>
 </Preview>
